Move tests to QuickTests
